requirements.txt
streamlit
pandas
requests
beautifulsoup4
newspaper3k
openpyxl
xlsxwriter
lxml